moving the iavi app onto the forms engine.  changing the registration and test sequences from *#blah#blah#blah#* to "prep register blah.." and "prep test blah...".
Updating the tests accordingly.

There is one test still failing (the test sequence test).  This has
something to do with how the test harness bootstraps all the apps and
the router being used to initiate the tree sequence not being the one
that is processing the script, which makes it not get the message.

The test sequence works in the httptester.