added the FOLLOW app...
...which uses DYNAMICALLY GENERATED MODELS (yes!) to hook it's little
claws into existing models. so long as you can resist looking inside
the pit of despair that is apps.follow.models, it provides an easy
way for Reporters and/or Connections to "follow" other models. of
course, "following" something doesn't actually do anything yet.

any model can become followable, but at the moment, i've limited it
to models that expose a __search__ classmethod, which is used by
apps.follow.app.App to search (sort of) for things to follow. for
example:

  ~>> I AM adammck
  <<~ Hello, Adam Mckaig!

  ~>> FOLLOW evan wheeler mej nyc
  <<~ You are now following Evan Wheeler, Mark Johnston, New York City

note the lack of delimiters, and loose search terms.
later, in the shell:

  >>> evan = Reporter.objects.get(pk=2)
  <Reporter: Evan Wheeler (ewheeler)>

  >>> evan.followers.all()
  [<FollowingReporter: <Reporter: Adam Mckaig (adammck)>>]

  >>> nyc = Location.objects.get(code="NYC")
  <Location: New York City (NYC)>

  >>> nyc.followers.all()
  [<FollowingLocation: <Reporter: Adam Mckaig (adammck)>>]

the same works in reverse, although it's a little more verbose, since
each following model is a separate foreign key. they could be collated
by a helper method, some time.

  >>> adam = Reporter.objects.get(pk=1)
  <Reporter: Adam Mckaig (adammck)>

  >>> adam.following_reporter_set.all()
  [<Reporter: Evan Wheeler (ewheeler)>, <Reporter: Mark Johnson (mej)>]